+++ sources/src/main/java/com/google/solutions/jitaccess/web/ApiResource.java
@@ -19,12 +19,12 @@
 // under the License.
 //
 
-package com.google.solutions.jitaccess.web.rest;
+package com.google.solutions.jitaccess.web;
 
-import com.google.common.base.Preconditions;
 import com.google.solutions.jitaccess.core.AccessDeniedException;
 import com.google.solutions.jitaccess.core.AccessException;
 import com.google.solutions.jitaccess.web.iap.IapPrincipal;
+import com.google.solutions.jitaccess.web.actions.*;
 import jakarta.enterprise.context.Dependent;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.*;
