+++ sources/src/main/java/com/google/solutions/jitaccess/web/actions/ListPeersAction.java
@@ -19,15 +19,14 @@
 // under the License.
 //
 
-package com.google.solutions.jitaccess.web.rest;
+package com.google.solutions.jitaccess.web.actions;
 
 import com.google.common.base.Preconditions;
 import com.google.solutions.jitaccess.core.AccessDeniedException;
 import com.google.solutions.jitaccess.core.AccessException;
 import com.google.solutions.jitaccess.core.Exceptions;
 import com.google.solutions.jitaccess.core.RoleBinding;
-import com.google.solutions.jitaccess.core.catalog.ActivationType;
-import com.google.solutions.jitaccess.core.catalog.Entitlement;
+import com.google.solutions.jitaccess.core.auth.UserEmail;
 import com.google.solutions.jitaccess.core.catalog.ProjectId;
 import com.google.solutions.jitaccess.core.catalog.project.MpaProjectRoleCatalog;
 import com.google.solutions.jitaccess.web.LogAdapter;
@@ -37,18 +36,16 @@ import jakarta.enterprise.context.Dependent;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 /**
- * List roles (within a project) that the user can activate.
+ * List peers that are qualified to approve the activation of a role.
  */
 @Dependent
-public class ListRolesAction extends AbstractAction {
+public class ListPeersAction extends AbstractAction {
   private final @NotNull MpaProjectRoleCatalog catalog;
 
-  public ListRolesAction(
+  public ListPeersAction(
     @NotNull LogAdapter logAdapter,
     @NotNull MpaProjectRoleCatalog catalog
   ) {
@@ -58,75 +55,51 @@ public class ListRolesAction extends AbstractAction {
 
   public @NotNull ResponseEntity execute(
     @NotNull IapPrincipal iapPrincipal,
-    @Nullable String projectIdString
+    @Nullable String projectIdString,
+    @Nullable String role
   ) throws AccessException {
     Preconditions.checkArgument(
       projectIdString != null && !projectIdString.trim().isEmpty(),
       "A projectId is required");
+    Preconditions.checkArgument(
+      role != null && !role.trim().isEmpty(),
+      "A role is required");
 
     var userContext = this.catalog.createContext(iapPrincipal.email());
+
     var projectId = new ProjectId(projectIdString);
+    var roleBinding = new RoleBinding(projectId, role);
 
     try {
-      var entitlements = this.catalog.listEntitlements(userContext, projectId);
+      var peers = this.catalog.listReviewers(
+        userContext,
+        new com.google.solutions.jitaccess.core.catalog.project.ProjectRole(roleBinding));
 
-      return new ResponseEntity(
-        entitlements.available()
-          .stream()
-          .map(ent -> new ResponseEntity.Item(
-            ent.id().roleBinding(),
-            ent.activationType(),
-            ent.status(),
-            ent.validity() != null ? ent.validity().end().getEpochSecond() : null))
-          .collect(Collectors.toList()),
-        entitlements.warnings());
+      assert !peers.contains(iapPrincipal.email());
+
+      return new ResponseEntity(peers);
     }
     catch (Exception e) {
       this.logAdapter
         .newErrorEntry(
-          LogEvents.API_LIST_ROLES,
-          String.format("Listing project roles failed: %s", Exceptions.getFullMessage(e)))
+          LogEvents.API_LIST_PEERS,
+          String.format("Listing peers failed: %s", Exceptions.getFullMessage(e)))
         .addLabels(le -> addLabels(le, e))
+        .addLabels(le -> addLabels(le, roleBinding))
         .addLabels(le -> addLabels(le, projectId))
         .write();
 
-      throw new AccessDeniedException("Listing project roles failed, see logs for details");
+      throw new AccessDeniedException("Listing peers failed, see logs for details");
     }
   }
 
+
   public static class ResponseEntity {
-    public final Set<String> warnings;
-    public final @NotNull List<Item> roles;
+    public final @NotNull Set<UserEmail> peers;
 
-    private ResponseEntity(
-      @NotNull List<Item> roles,
-      Set<String> warnings
-    ) {
-      Preconditions.checkNotNull(roles, "roles");
-
-      this.warnings = warnings;
-      this.roles = roles;
-    }
-
-    public static class Item {
-      public final @NotNull RoleBinding roleBinding;
-      public final ActivationType activationType;
-      public final Entitlement.Status status;
-      public final Long /* optional */ validUntil;
-
-      public Item(
-        @NotNull RoleBinding roleBinding,
-        ActivationType activationType,
-        Entitlement.Status status,
-        Long validUntil) {
-
-        Preconditions.checkNotNull(roleBinding, "roleBinding");
-
-        this.roleBinding = roleBinding;
-        this.activationType = activationType;
-        this.status = status;
-        this.validUntil = validUntil;
-      }
+    private ResponseEntity(@NotNull Set<UserEmail> peers) {
+      Preconditions.checkNotNull(peers);
+      this.peers = peers;
     }
   }
 }
