+++ sources/src/main/java/com/google/solutions/jitaccess/web/actions/ListRolesAction.java
@@ -19,14 +19,15 @@
 // under the License.
 //
 
-package com.google.solutions.jitaccess.web.rest;
+package com.google.solutions.jitaccess.web.actions;
 
 import com.google.common.base.Preconditions;
 import com.google.solutions.jitaccess.core.AccessDeniedException;
 import com.google.solutions.jitaccess.core.AccessException;
 import com.google.solutions.jitaccess.core.Exceptions;
 import com.google.solutions.jitaccess.core.RoleBinding;
-import com.google.solutions.jitaccess.core.auth.UserEmail;
+import com.google.solutions.jitaccess.core.catalog.ActivationType;
+import com.google.solutions.jitaccess.core.catalog.Entitlement;
 import com.google.solutions.jitaccess.core.catalog.ProjectId;
 import com.google.solutions.jitaccess.core.catalog.project.MpaProjectRoleCatalog;
 import com.google.solutions.jitaccess.web.LogAdapter;
@@ -36,16 +37,18 @@ import jakarta.enterprise.context.Dependent;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
- * List peers that are qualified to approve the activation of a role.
+ * List roles (within a project) that the user can activate.
  */
 @Dependent
-public class ListPeersAction extends AbstractAction {
+public class ListRolesAction extends AbstractAction {
   private final @NotNull MpaProjectRoleCatalog catalog;
 
-  public ListPeersAction(
+  public ListRolesAction(
     @NotNull LogAdapter logAdapter,
     @NotNull MpaProjectRoleCatalog catalog
   ) {
@@ -55,51 +58,75 @@ public class ListPeersAction extends AbstractAction {
 
   public @NotNull ResponseEntity execute(
     @NotNull IapPrincipal iapPrincipal,
-    @Nullable String projectIdString,
-    @Nullable String role
+    @Nullable String projectIdString
   ) throws AccessException {
     Preconditions.checkArgument(
       projectIdString != null && !projectIdString.trim().isEmpty(),
       "A projectId is required");
-    Preconditions.checkArgument(
-      role != null && !role.trim().isEmpty(),
-      "A role is required");
 
     var userContext = this.catalog.createContext(iapPrincipal.email());
-
     var projectId = new ProjectId(projectIdString);
-    var roleBinding = new RoleBinding(projectId, role);
 
     try {
-      var peers = this.catalog.listReviewers(
-        userContext,
-        new com.google.solutions.jitaccess.core.catalog.project.ProjectRole(roleBinding));
+      var entitlements = this.catalog.listEntitlements(userContext, projectId);
 
-      assert !peers.contains(iapPrincipal.email());
-
-      return new ResponseEntity(peers);
+      return new ResponseEntity(
+        entitlements.available()
+          .stream()
+          .map(ent -> new ResponseEntity.Item(
+            ent.id().roleBinding(),
+            ent.activationType(),
+            ent.status(),
+            ent.validity() != null ? ent.validity().end().getEpochSecond() : null))
+          .collect(Collectors.toList()),
+        entitlements.warnings());
     }
     catch (Exception e) {
       this.logAdapter
         .newErrorEntry(
-          LogEvents.API_LIST_PEERS,
-          String.format("Listing peers failed: %s", Exceptions.getFullMessage(e)))
+          LogEvents.API_LIST_ROLES,
+          String.format("Listing project roles failed: %s", Exceptions.getFullMessage(e)))
         .addLabels(le -> addLabels(le, e))
-        .addLabels(le -> addLabels(le, roleBinding))
         .addLabels(le -> addLabels(le, projectId))
         .write();
 
-      throw new AccessDeniedException("Listing peers failed, see logs for details");
+      throw new AccessDeniedException("Listing project roles failed, see logs for details");
     }
   }
 
-
   public static class ResponseEntity {
-    public final @NotNull Set<UserEmail> peers;
+    public final Set<String> warnings;
+    public final @NotNull List<Item> roles;
 
-    private ResponseEntity(@NotNull Set<UserEmail> peers) {
-      Preconditions.checkNotNull(peers);
-      this.peers = peers;
+    private ResponseEntity(
+      @NotNull List<Item> roles,
+      Set<String> warnings
+    ) {
+      Preconditions.checkNotNull(roles, "roles");
+
+      this.warnings = warnings;
+      this.roles = roles;
+    }
+
+    public static class Item {
+      public final @NotNull RoleBinding roleBinding;
+      public final ActivationType activationType;
+      public final Entitlement.Status status;
+      public final Long /* optional */ validUntil;
+
+      public Item(
+        @NotNull RoleBinding roleBinding,
+        ActivationType activationType,
+        Entitlement.Status status,
+        Long validUntil) {
+
+        Preconditions.checkNotNull(roleBinding, "roleBinding");
+
+        this.roleBinding = roleBinding;
+        this.activationType = activationType;
+        this.status = status;
+        this.validUntil = validUntil;
+      }
     }
   }
 }
