+++ sources/src/test/java/com/google/solutions/jitaccess/web/RestDispatcher.java
@@ -19,10 +19,11 @@
 // under the License.
 //
 
-package com.google.solutions.jitaccess.web.rest;
+package com.google.solutions.jitaccess.web;
 
 import com.google.gson.Gson;
 import com.google.solutions.jitaccess.core.auth.UserEmail;
+import com.google.solutions.jitaccess.web.ExceptionMappers;
 import com.google.solutions.jitaccess.web.iap.DeviceInfo;
 import com.google.solutions.jitaccess.web.iap.IapPrincipal;
 import jakarta.ws.rs.core.MediaType;
